feat: implement Leveled Vector Store, tiered HNSW index (frozen & delta layers), python SDK client, and optimization benchmarks
- Implemented LeveledStore in kvstore/vector/leveled_store.go along with delta.go and frozen.go to support tiered vector indexing.
- Added comprehensive unit and benchmark tests for the leveled store (leveled_store_test.go, leveled_bench_test.go).
- Created Python SDK client (molten_client.py), async benchmark runner, and bio loader in sdk/python.
- Added real-world bioinformatics benchmark, product quantization / LSH benchmarks, and EF search tests.
- Updated main server executable, CLI benchmark utility, and HNSW optimization roadmap.

+++ kvstore/vector/vector_arena.go
@@ -7,14 +7,14 @@ package vector
 type VectorArena struct {
 	data        []float32 // Единый плоский массив координат
 	dim         int       // Размерность вектора (например, 128)
-	freeOffsets []uint32  // Стек свободных смещений (Free List)
+	freeOffsets []uint64  // Стек свободных смещений (Free List)
 }
 
 func NewVectorArena(dim int, initialCapacity int) *VectorArena {
 	return &VectorArena{
 		data:        make([]float32, 0, initialCapacity*dim),
 		dim:         dim,
-		freeOffsets: make([]uint32, 0, 64),
+		freeOffsets: make([]uint64, 0, 64),
 	}
 }
 
@@ -22,7 +22,7 @@ func NewVectorArena(dim int, initialCapacity int) *VectorArena {
 //
 // Паникует если len(vec) != dim — несоответствие размерности
 // ломает весь layout арены (тихая порча данных).
-func (va *VectorArena) Allocate(vec []float32) uint32 {
+func (va *VectorArena) Allocate(vec []float32) uint64 {
 	if len(vec) != va.dim {
 		panic("VectorArena.Allocate: dimension mismatch")
 	}
@@ -31,11 +31,11 @@ func (va *VectorArena) Allocate(vec []float32) uint32 {
 	if nFree > 0 {
 		offset := va.freeOffsets[nFree-1]
 		va.freeOffsets = va.freeOffsets[:nFree-1]
-		copy(va.data[offset:offset+uint32(va.dim)], vec)
+		copy(va.data[offset:offset+uint64(va.dim)], vec)
 		return offset
 	}
 
-	offset := uint32(len(va.data))
+	offset := uint64(len(va.data))
 	va.data = append(va.data, vec...)
 	return offset
 }
@@ -44,11 +44,11 @@ func (va *VectorArena) Allocate(vec []float32) uint32 {
 //
 // Контракт: вызывающий код гарантирует, что offset не освобождается дважды.
 // Double-free — ошибка программиста, не проверяется ради O(1).
-func (va *VectorArena) Free(offset uint32) {
+func (va *VectorArena) Free(offset uint64) {
 	va.freeOffsets = append(va.freeOffsets, offset)
 }
 
 // Get возвращает быстрый слайс-взгляд на вектор по смещению
-func (va *VectorArena) Get(offset uint32) []float32 {
-	return va.data[offset : offset+uint32(va.dim)]
+func (va *VectorArena) Get(offset uint64) []float32 {
+	return va.data[offset : offset+uint64(va.dim)]
 }
